♻️ refactor(delta-plugin): 重构插件代码结构，实现模块化分离
- 将核心数据服务类 `DeltaService` 及相关常量、映射移至新建的 `data_source.py` 模块
- 将指令处理逻辑及主程序入口移至新建的 `main.py` 模块
- 简化 `__init__.py` 文件，仅保留插件元数据、服务实例化及启动/关闭生命周期管理
- 提升代码可读性与可维护性，明确职责分离
